pyproject.toml
[project]
name = "ai-engineering-workshop"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "tiktoken>=0.12.0",
  "pypdf>=6.9.2",
  "pandas>=2.3.3",
  "ipykernel>=7.2.0",
  "fastapi[standard]",
  "langchain",
  "langgraph",
  "langsmith",
  "langchain-openai",
  "python-dotenv",
  "pydantic-settings",
  "psycopg[binary,pool]",
  "langgraph-checkpoint-postgres",
]

[project.optional-dependencies]
rag = [
  "faiss-cpu>=1.13.2",
  "langchain-text-splitters>=1.1.1",
]

ml = [
  "scikit-learn>=1.7.2",
  "xgboost>=3.1.2",
  "lightgbm>=4.6.0",
  "catboost>=1.2.8",
  "imbalanced-learn>=0.14.1",
  "optuna>=4.6.0",
  "shap>=0.49.1",
]

dl = [
  "tensorflow>=2.20.0",
  "tensorboard>=2.20.0",
  "keras-tuner>=1.4.8",
]

rl = [
  "gymnasium[classic-control]>=1.2.3",
  "stable-baselines3>=2.7.1",
]

scraping = [
  "scrapy>=2.13.3",
  "selenium>=4.38.0",
]

notebooks = [
  "ipykernel>=6.30.1",
  "notebook>=7.4.7",
  "nbformat>=5.10.4",
]

ops = [
  "mlflow>=3.8.1",
  "zenml[local,server]>=0.93.1",
]
